@@ -1,4 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ReptilesService } from '../services/reptiles.service';
+import { Reptile } from '../models/reptile.model';
+import { Subscription } from 'rxjs/Subscription';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-reptile-list',
@@ -7,9 +11,34 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ReptileListComponent implements OnInit {
 
-  constructor() { }
+  reptiles: Reptile[];
+  reptilesSubscription: Subscription;
+
+  constructor(private reptilesService: ReptilesService, private router: Router) {}
 
   ngOnInit() {
+    this.reptilesSubscription = this.reptilesService.reptilesSubject.subscribe(
+      (reptiles: Reptile[]) => {
+        this.reptiles = reptiles;
+      }
+    );
+    this.reptilesService.emitReptiles();
+  }
+
+  onNewReptile() {
+    this.router.navigate(['/reptiles', 'new']);
+  }
+
+  onDeleteReptile(reptile: Reptile) {
+    this.reptilesService.removeReptile(reptile);
+  }
+
+  onViewReptile(id: number) {
+    this.router.navigate(['/reptiles', 'view', id]);
+  }
+
+  ngOnDestroy() {
+    this.reptilesSubscription.unsubscribe();
   }
 
 }
